requirements.txt
asgiref==3.5.2
Django==4.1.3
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==22.1
django-mathfilters==1.0.0
django-suit==0.2.25
Pillow==9.3.0
python-decouple==3.6
sqlparse==0.4.3
